Fix folder names
The folder name was sometimes missing a few characters at the end due to
a misinterpretation of the str.rstrip() method. The intention was to
remove the trailing string ".git", gut the old code was removing all
occurrences of the characters ".", "g", "i" and "t" at the end of the
folder name. Replaced that part of the code with an explicit test using
str.endswith().